+++ zapisy/apps/api/rest/v1/serializers.py
@@ -3,6 +3,8 @@ from rest_framework import serializers
 
 from apps.enrollment.courses.models.classroom import Classroom
 from apps.enrollment.courses.models.semester import Semester
+from apps.offer.desiderata.models import Desiderata, DesiderataOther
+from apps.schedule.models.specialreservation import SpecialReservation
 from apps.users.models import Employee
 
 
@@ -36,3 +38,21 @@ class EmployeeSerializer(serializers.ModelSerializer):
     class Meta:
         model = Employee
         fields = ('id', 'user', 'consultations', 'homepage', 'room')
+
+
+class DesiderataSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Desiderata
+        fields = '__all__'
+
+
+class DesiderataOtherSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = DesiderataOther
+        fields = '__all__'
+
+
+class SpecialReservationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = SpecialReservation
+        fields = '__all__'