+++ zapisy/apps/enrollment/courses/models/classroom.py
@@ -61,8 +61,9 @@ class Classroom(models.Model):
 
     @classmethod
     def get_terms_in_day(cls, date, ajax=False):
-        from apps.schedule.models import Term as EventTerm
-        from apps.enrollment.courses.models import Semester, Term, Freeday, ChangedDay
+        from apps.schedule.models.term import Term as EventTerm
+        from apps.enrollment.courses.models.semester import Semester, Freeday, ChangedDay
+        from apps.enrollment.courses.models.term import Term
 
         rooms = cls.get_in_institute(reservation=True)