+++ zapisy/apps/schedule/management/commands/importExams.py
@@ -2,17 +2,15 @@ import datetime
 
 from django.contrib.auth.models import User
 from django.core.management import BaseCommand
-from django.db.models import Q
 
-from apps.enrollment.courses.models import Semester, Freeday, ChangedDay, Classroom, Course
-from apps.enrollment.courses.models.term import Term as T
-from apps.schedule.models import Term, Event
+from apps.enrollment.courses.models.classroom import Classroom
+from apps.enrollment.courses.models.course import Course
+from apps.schedule.models.term import Term
+from apps.schedule.models.event import Event
 
 
 class Command(BaseCommand):
-
     def handle(self, *args, **options):
-
         s4 = Classroom.objects.get(number='4')
         s5 = Classroom.objects.get(number='5')
         s7 = Classroom.objects.get(number='7')