+++ zapisy/apps/users/models.py
@@ -132,15 +132,6 @@ class Employee(BaseUser):
         verbose_name="Status")
     title = models.CharField(max_length=20, verbose_name="tytuł naukowy", null=True, blank=True)
 
-    def make_preferences(self):
-        from apps.offer.preferences.models import Preference
-
-        Preference.make_preferences(self)
-
-    def get_preferences(self):
-        from apps.offer.preferences.models import Preference
-        return Preference.for_employee(self)
-
     def has_privileges_for_group(self, group_id):
         """
         Method used to verify whether user is allowed to create a poll for certain group
@@ -191,7 +182,8 @@ class Employee(BaseUser):
 
     @staticmethod
     def get_all_groups_in_semester(user_id):
-        from apps.enrollment.courses.models import Group, Semester
+        from apps.enrollment.courses.models.group import Group
+        from apps.enrollment.courses.models.semester import Semester
 
         user = User.objects.get(id=user_id)
         semester = Semester.get_default_semester()
@@ -357,7 +349,8 @@ class Student(BaseUser):
         return list(frozenset(records_list))
 
     def get_points(self, semester=None):
-        from apps.enrollment.courses.models import Semester, StudentPointsView
+        from apps.enrollment.courses.models.semester import Semester
+        from apps.enrollment.courses.models.points import StudentPointsView
         from apps.enrollment.records.models import Record
         if not semester:
             semester = Semester.objects.get_next()
@@ -372,7 +365,8 @@ class Student(BaseUser):
         return StudentPointsView.get_points_for_entities(self, records)
 
     def get_points_with_course(self, course, semester=None):
-        from apps.enrollment.courses.models import Semester, StudentPointsView
+        from apps.enrollment.courses.models.semester import Semester
+        from apps.enrollment.courses.models.points import StudentPointsView
         from apps.enrollment.records.models import Record
         if not semester:
             semester = Semester.objects.get_next()
@@ -404,21 +398,24 @@ class Student(BaseUser):
         return cls.objects.filter(status=0)
 
     @staticmethod
-    def get_list(begin='All'):
+    def get_list(begin='All', restrict_list_consent=True):
         def next_char(begin):
             try:
                 return chr(ord(begin) + 1)
             except ValueError:
                 return chr(90)
+
+        qs = Student.objects.filter(status=0)
+        if restrict_list_consent:
+            qs = qs.filter(consent__granted=True)
         if begin == 'Z':
-            return Student.objects.filter(status=0, user__last_name__gte=begin).\
+            return qs.filter(user__last_name__gte=begin).\
                 select_related().order_by('user__last_name', 'user__first_name')
         elif begin == 'All':
-            return Student.objects.filter(status=0).\
-                select_related().order_by('user__last_name', 'user__first_name')
+            return qs.select_related().order_by('user__last_name', 'user__first_name')
         else:
             end = next_char(begin)
-            return Student.objects.filter(status=0, user__last_name__range=(begin, end)).\
+            return qs.filter(user__last_name__range=(begin, end)).\
                 select_related().order_by('user__last_name', 'user__first_name')
 
     @staticmethod
@@ -488,6 +485,12 @@ class Student(BaseUser):
 
     def is_fresh_student(self):
         return True
+
+    def consent_answered(self):
+        return hasattr(self, 'consent')
+
+    def consent_granted(self):
+        return self.consent_answered() and self.consent.granted
 
     class Meta:
         verbose_name = 'student'
@@ -725,3 +728,18 @@ class OpeningTimesView(models.Model):
 
     class Meta:
         app_label = 'users'
+
+
+class PersonalDataConsent(models.Model):
+    """
+        Model przechowuje zgody dotyczące udostępniania danych osobowych studentów
+    """
+    student = models.OneToOneField(Student, related_name='consent', on_delete=models.CASCADE)
+    granted = models.NullBooleanField(verbose_name="zgoda udzielona")
+
+    class Meta:
+        verbose_name = 'Zgoda na udostępnianie danych osobowych'
+        verbose_name_plural = 'Zgody na udostępnianie danych osobowych'
+
+    def __str__(self):
+        return f"{self.student.get_full_name()}: {self.granted}"