+++ zapisy/test_app/tests.py
@@ -3,8 +3,6 @@
 from django.test import LiveServerTestCase
 
 from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.firefox.webdriver import WebDriver
 from selenium.webdriver.support.ui import Select
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
@@ -16,8 +14,13 @@ from selenium.common.exceptions import (
 )
 
 from django.contrib.auth.models import User
-from apps.users.models import Employee, Student
-from apps.enrollment.courses.models import Semester, CourseEntity, Course, Type, Group, Term, Classroom
+from apps.users.models import Employee, Student, PersonalDataConsent
+from apps.enrollment.courses.models.semester import Semester
+from apps.enrollment.courses.models.course import CourseEntity, Course
+from apps.enrollment.courses.models.course_type import Type
+from apps.enrollment.courses.models.group import Group
+from apps.enrollment.courses.models.term import Term
+from apps.enrollment.courses.models.classroom import Classroom
 from apps.offer.vote.models import SystemState
 
 import os
@@ -112,7 +115,14 @@ class NewSemesterTests(SeleniumTestCase):
             user=user_student3, matricula='333333')
         self.student4 = Student.objects.create(
             user=user_student4, matricula='444444')
-
+        PersonalDataConsent.objects.update_or_create(student=self.student1,
+                                                     defaults={'granted': True})
+        PersonalDataConsent.objects.update_or_create(student=self.student2,
+                                                     defaults={'granted': True})
+        PersonalDataConsent.objects.update_or_create(student=self.student3,
+                                                     defaults={'granted': True})
+        PersonalDataConsent.objects.update_or_create(student=self.student4,
+                                                     defaults={'granted': True})
         self.course_type = Type.objects.create(name='Informatyczny')
         for i in range(1, 6):
             CourseEntity.objects.create(